pkg/sysconfwatch: add non-linux variant to allow microshift binary to build
The binary itself might not run as expected on Apple or other non-linux
platforms, but we should avoid adding linux specific code without
alternatives for other systems/platforms.

As a bonus, this allow to build microshift binary on M1 MacBook.
